Westwood PAK support
Now I gotta note, that this was a very extremely messy file format for the purpose it was set up for. I really wonder what the guys at Westwood were thinking.

What is also notable is that there is not a real way to recognize Westwood PAK file properly, so this can already cause some trouble, but JCR6 tries the best it can with this driver. Good part... it can be turned off to win performance *AND* stability.

In normal usage it *SHOULD* work fine. This driver has been tested with Lands Of Lore - The Throne of Chaos, and the results I got there, were satisfactory.